abstract connect function
