+++ library/observables-connect/src/accounts/index.ts
@@ -1,6 +1,7 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
+export * from './connect'
 export * from './get'
 export * from './local'
 export * from './observables'
