+++ library/react-connect-kit/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@w3ux/react-connect-kit-source",
   "license": "GPL-3.0-only",
-  "version": "3.2.19",
+  "version": "3.2.20",
   "type": "module",
   "scripts": {
     "clear": "rm -rf node_modules dist tsconfig.tsbuildinfo",
@@ -10,7 +10,7 @@
   "dependencies": {
     "@w3ux/extension-assets": "^2.2.0",
     "@w3ux/hooks": "^2.1.0",
-    "@w3ux/observables-connect": "0.9.26",
+    "@w3ux/observables-connect": "0.9.27",
     "@w3ux/utils": "^2.0.8"
   },
   "devDependencies": {
